Don't scroll automatically if end of list appears in middle of table
The first request (from the specifications) happens to be the third jedi
from the end of the list, so the final jedi appeared in the third row.
In the previous implementation, that row was auto-scrolled to the bottom
and the table was filled at the top. This commit removes that behaviour
after staltz's comments in https://github.com/staltz/flux-challenge/pull/58.